requirement.txt
streamlit==1.28.0
google-generativeai==0.3.0
chromadb==0.4.15
sentence-transformers==2.2.2
requests==2.31.0
python-dotenv==1.0.0
pillow==10.0.1
pandas==2.1.0
tqdm==4.66.1